LF-3803 extract repeated breakpoint logic to hook
